chore: Rename Context types file
